requirements.txt
flask
flask-socketio
# simple-websocket
discord.py
pathvalidate
reactionmenu
simple-websocket
opencv-python-headless